Add updateUrl in app build instead of too late in the installer build
